+++ include/network/session/SessionManager.h
@@ -13,16 +13,16 @@ class SessionManager
 public:
     static SessionManager &Instance();
 
-    std::shared_ptr<Session> CreateSessionWithId(uint64_t id);
+    std::shared_ptr<Session> CreateSessionWithId(uint32_t id);
 
-    std::shared_ptr<Session> GetSession(uint64_t id);
+    std::shared_ptr<Session> GetSession(uint32_t id);
 
-    void RemoveSession(uint64_t id);
+    void RemoveSession(uint32_t id);
 
     void CheckTimeout();
 
     // 🔥 新增：通过 PlayerId 查找并踢掉旧 Session
-    void KickPlayer(uint64_t playerId, uint64_t excludeSessionId, const std::string &reason);
+    void KickPlayer(uint64_t playerId, uint32_t excludeSessionId, const std::string &reason);
 
     // 在 Session 绑定 Player 时调用，建立映射
     void BindPlayerToSession(uint64_t playerId, std::shared_ptr<Session> session);
@@ -38,13 +38,13 @@ private:
     struct Bucket
     {
         std::mutex mutex;
-        std::unordered_map<uint64_t, std::shared_ptr<Session>> sessions;
+        std::unordered_map<uint32_t, std::shared_ptr<Session>> sessions;
     };
 
     Bucket buckets_[BUCKET_COUNT];
 
 private:
-    size_t GetBucketIndex(uint64_t id) const
+    size_t GetBucketIndex(uint32_t id) const
     {
         return id % BUCKET_COUNT;
     }
